test: move fixture setup_dummy_sensors from test_reporting.py to conftest.py
